% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{abspath}
\alias{abspath}
\title{Turn path into an absolute path}
\usage{
abspath(path)
}
\arguments{
\item{path}{a relative path}
}
\value{
an absolute path
}
\description{
Turn path into an aboslute path. The file or folder does not have to exists.
}
\details{
The \code{tools::file_path_as_absolute} will turn a path into an aboslute path but only if
the path exists. This makes it unusable in certain cases, such as when a desired output path
needs to be passed to an external call.
}
\examples{
abspath("foo.txt") # equal to file.path(getwd(), "foo.txt")
abspath("~/foo/bar/")

}
