% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{all.files.exist}
\alias{all.files.exist}
\title{Check existence of files}
\usage{
all.files.exist(x)
}
\arguments{
\item{x}{a vector or list of files.}
}
\value{
a logical value indicating if all files exist.
}
\description{
This function checks the existence of all files stored in a vector or a list.
}
\details{
\code{all.files.exist} is a simple shorthand around \code{\link[base:files]{base::file.exists}} wrapped with an additional
\code{\link[base:all]{base::all}} to return a single value. This makes this function useful when checking that all
files required for or created by some function exist.
}
\examples{
files = c(tempfile(), tempfile())
all.files.exist(files) # FALSE
file.create(files)
all.files.exist(files) # TRUE

}
