% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamtagregex.r
\name{bamtagregex}
\alias{bamtagregex}
\title{Change sam/bam tag values}
\usage{
bamtagregex(input, output, tag, pattern, replace)
}
\arguments{
\item{input}{an input sam/bam file}

\item{output}{an output sam/bam file}

\item{tag}{a tag to modify}

\item{pattern}{a regex pattern}

\item{replace}{a string to replace matched pattern}
}
\description{
Change tag value in sam/bam file using regular expression.
}
\details{
This code is a thin wrapper over the Python script \code{bamtagregex.py} found
in the package's \code{inst} folder.

The \code{bamtagregex.py} uses the amazing \code{pysam} package to manipulate the sam/bam files.
The Python's \code{pysam} package is powered by the \code{htslib} library. the same library
that powers \code{samtools}.
In R, there is a similar package to \code{pysam}, the \code{Rsamtools} found in bioconductor. But this
library reads the whole file at once (instead of using memory efficient iterators), which
makes working with a large bam files less than ideal.
}
\examples{
\dontrun{
# Replace the end of 10x barcode from "-1" to "-baz"
bamtagregex("foo.bam", "bar.bam", "CB", "-1", "-baz")
}
}
