% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamtofastq.r
\name{bamtofastq}
\alias{bamtofastq}
\title{BAM to fastq}
\usage{
bamtofastq(bam, outdir, nthreads = 4, remake = FALSE)
}
\arguments{
\item{bam}{an input BAM file that was previously mapped using the Cellranger software}

\item{outdir}{a non-existing output directory, see details}

\item{nthreads}{a number of threads to run on}

\item{remake}{whether to remake the output if it already exists}
}
\description{
Split a BAM aligned using the Cellranger software into individual \code{.fastq} files.
Typically this is done to re-map and re-analyze processed data.
}
\details{
To mark a succesful run, the function creates a \code{completed} status file in the output directory.
If such file exists and \code{remake=TRUE}, the output directory is deleted and the output is
recreated.

However, if the output directory \code{outdir} exists, but the status file does not, an error
is reported. This is to prevent cases when an incorrect existing \code{outdir} would be delted.
}
\seealso{
\link{cellranger_count} to re-map and re-analyze \code{.fastq} files.
}
