% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger.r
\name{cellranger_count}
\alias{cellranger_count}
\title{Estimate scRNAseq Gene Expression with Cellranger Count}
\usage{
cellranger_count(
  id,
  fastqdir,
  refdir,
  outdir = ".",
  chemistry = "auto",
  nthreads = 4,
  sample = NULL,
  remake = FALSE
)
}
\arguments{
\item{id}{an id that identifies the sample, this will be embedded in the read-groups (\code{RG} tag)
of the header and reads of the bam-file.}

\item{fastqdir}{a dir with fastq files prepared with \link{bamtofastq}}

\item{refdir}{a directory with the reference files created by the \link{cellranger_mkref}}

\item{outdir}{a non-existing output directory, see details}

\item{chemistry}{\strong{optional} a 10X chemistry, use only when the autodetection is failing}

\item{nthreads}{\strong{optional} a number of threads to run on}

\item{sample}{\strong{optional} a file prefix to specify which sample to select,
required if fastqdir contains fastq files for multiple samples}

\item{remake}{\strong{optional} remake the output if it already exists}
}
\description{
Perform mapping, cell estimation, filtering and gene expression analysis using the Cellranger
software.
}
\details{
To mark a succesful run, the function creates a \code{completed} status file in the output directory.
If such file exists and \code{remake=TRUE}, the output directory is deleted and the output is
recreated.

However, if the output directory \code{outdir} exists, but the status file does not, an error
is reported. This is to prevent cases when an incorrect existing \code{outdir} would be delted.
}
\seealso{
\link{cellranger_mkref} to create the required reference genome directory (\code{refdir})
\link{bamtofastq} to transform mapped BAM file back into fasta files for remapping
}
