% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger.r
\name{cellranger_mkref}
\alias{cellranger_mkref}
\title{Generate Cellranger Transcriptome Reference Genome}
\usage{
cellranger_mkref(reference, annotation, outdir, nthreads = 4, remake = FALSE)
}
\arguments{
\item{reference}{a reference genome fasta file (e.g., \code{.fna})}

\item{annotation}{a GTF (\code{.gtf}) annotation file associated with the reference genome}

\item{outdir}{a non-existing output directory, see details}

\item{nthreads}{\strong{optional} a number of threads to run on}

\item{remake}{\strong{optional} remake the output if it already exists}
}
\description{
Run \verb{cellranger mkref} and generate the Cellranger Transcriptome Reference Genome.
}
\details{
The \emph{reference} which must include associated index file (\code{fai}) in the same directory.

To mark a succesful run, the function creates a \code{completed} status file in the output directory.
If such file exists and \code{remake=TRUE}, the output directory is deleted and the output is
recreated.

However, if the output directory \code{outdir} exists, but the status file does not, an error
is reported. This is to prevent cases when an incorrect existing \code{outdir} would be delted.
}
\seealso{
\link{cellranger_count} to use this reference directory to perform an expression analysis
\link{bamtofastq} to transform mapped BAM file back into fasta files for remapping
}
