% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{corename}
\alias{corename}
\title{Get core part of the file name}
\usage{
corename(path)
}
\arguments{
\item{path}{a path to file}
}
\value{
a core part of the file name
}
\description{
This is similar to the base function \code{basename}. The difference
is that this also removes the file type (or extension) and thus
works as a shorthand for many file operations.
}
\examples{
corename("/my/path/my.file.ext") # my.file

}
