% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest_subset.r
\name{densest_subset}
\alias{densest_subset}
\title{Find the densest subset of a matrix}
\usage{
densest_subset(x, empty = NA, steps = 100, density = 1)
}
\arguments{
\item{x}{a matrix with unknown values}

\item{empty}{the unknown value if different from NA; such as \code{0} for a numeric matrix or \code{N},
\code{empty} or \code{unknown} for character matrix}

\item{steps}{the number of steps}

\item{density}{a desired density}
}
\value{
a list containing the densest submatrix, vectors of deleted columns and a vector of
densities  at each step of the algorithm
}
\description{
Given a matrix with unknown values, find a densest subset of matrix.
}
\details{
This is a stepwise filtering algorithm that deletes the least represented columns and/or rows
to get a matrix subset with the least amount of unknown values until an optimum is reached.

In a lot of cases, an optimum would be a single vector. For this reasons, the algorithm
is limited both by the number of steps and by a desired density.
}
