% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{empty_to_na}
\alias{empty_to_na}
\title{Convert a value to NA}
\usage{
empty_to_na(x, empty)
}
\arguments{
\item{x}{a vector, matrix or a data frame}

\item{empty}{a non-standard unknown value}
}
\value{
an object where a non-standard unknown value is replaced by \code{NA}
}
\description{
Convert a value considered to be an unknown value to the standard \code{NA}
}
\examples{
x = c("foo", "bar", "unknown")
phyloRNA:::empty_to_na(x, "unknown")
}
