% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.r
\name{fasta}
\alias{fasta}
\title{Transform matrix into a fasta format}
\usage{
fasta(data, margin = 2, file = NULL)
}
\arguments{
\item{data}{a character matrix}

\item{margin}{\strong{optional} whether rows (1) or columns (2) should be concatenated}

\item{file}{\strong{optional} write the fasta file into the specified file instead}
}
\value{
a character vector
}
\description{
Concatenate a character matrix rows or colums into an alignment in a fasta format.
}
\details{
Each cell of the character matrix must have length of the string exactly 1, otherwise
different size of the sequences is produced.
}
