% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.r
\name{fasta2table}
\alias{fasta2table}
\title{Convert a fasta into table}
\usage{
fasta2table(fasta, margin = 1)
}
\arguments{
\item{fasta}{a named vector of sequences, such as from \code{read_fasta}}

\item{margin}{\strong{optional} whether rows sequences should be in rows (1) or columns (2)}
}
\value{
a matrix of fasta sequences
}
\description{
Convert a fasta, named vector, into a matrix by splitting up sequences into bases/SNV/CNVs.
}
\examples{
fasta = c("ac", "bd")
mat = fasta2table(fasta)
identical(fasta, c("ac","bd"))

}
\seealso{
\link{write_fasta} to write fasta file,
\link{read_fasta} to read fasta file,
\link{table2fasta} to transform table into a fasta file
}
