% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gtf.r
\name{filter_gtf}
\alias{filter_gtf}
\title{Filter GTF file according to index file}
\usage{
filter_gtf(gtf, fai, output = NULL)
}
\arguments{
\item{gtf}{an input GTF file}

\item{fai}{an input fasta index file}

\item{output}{\strong{optional} an output file. If not specified, output is derived from the fasta
index file.}
}
\description{
Fitler an annotation GTF file according to fasta index file.
}
\details{
This might be useful when filtering a GTF so that only chromosomes or scaffolds of interest
are present.

This code is a thin wrapper over the Python script \code{filter_gtf.py} found in the package's \code{inst}
folder.
}
