% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_mutect2.r
\name{gatk_mutect2}
\alias{gatk_mutect2}
\alias{gatk_Mutect2}
\title{Mapping to the Mutect2}
\usage{
gatk_Mutect2(
  bam,
  reference,
  vcf,
  normal = NULL,
  pon = NULL,
  germline = NULL,
  ps = NULL,
  remake = FALSE
)
}
\arguments{
\item{bam}{one or more sam/bam files to where mutations are to be called and optionally
matched normal samples when \code{Mutect2} is called in the "Tumor with Matched Normal" mode}

\item{reference}{a reference fasta (\code{.fas}) file to which the sam/bam file was mapped}

\item{vcf}{an output \code{vcf} file with called mutations}

\item{normal}{\strong{optional} sample names (\code{RG} sam/bam tag) of the matched normal samples}

\item{pon}{\strong{optional} the Panel of Normals (see \code{\link{gatk_make_pon}})}

\item{germline}{\strong{optional} a vcf file of germline population contaning allelic fractions}

\item{ps}{\strong{optional} a parameter string passed to \verb{gatk Mutect2} call.}

\item{remake}{\strong{optional} remake the output if it already exists}
}
\description{
Mapping to the Genome Analysis ToolKit (GATK) Mutect2 somatic mutation caller.
}
\details{
Somatic mutations include short nucleid variations (SNV) and insertion-deletions (inserts).
}
\seealso{
\link{GATK} a GATK binding
\link{GATKR6} a GATK binding in the form of R6 class
\link{gatk_snv} a convenience function simplifying SNV calling with \code{gatk_Mutect2}
\link{gatk_prepare} and \link{gatk_make_pon} for other convenience functions utilizing the GATK calls
}
