% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_pon.r
\name{gatk_pon}
\alias{gatk_pon}
\alias{gatk_make_pon}
\alias{gatk_GenomicDBIImport}
\alias{gatk_CreateSomaticPanelOfNormals}
\title{Make a Panel of Normals using the GATK toolkit}
\usage{
gatk_make_pon(bam, reference, vcf, outdir = "pon", remake = FALSE)

gatk_GenomicDBIImport(vcf, reference, database, remake = remake)

gatk_CreateSomaticPanelOfNormals(database, reference, vcf, remake = FALSE)
}
\arguments{
\item{bam}{one or more input sam/bam files}

\item{reference}{a reference genome that was used during mapping of the sam/bam files}

\item{vcf}{a variant call file, either one or more vcf to be merged into a database
(for \code{gatk_GenomicDBIImport}) or an output vcf with Panel of Normals (for \code{gatk_make_pon} and
\code{gatk_CreateSomaticPanelOfNormals})}

\item{outdir}{\strong{optional} outdir an output directory for intermediate files}

\item{remake}{\strong{optional} remake the output if it already exists}

\item{database}{a name of the database file created by \code{gatk_GenomicDBIImport} and used in
\code{gatk_CreateSomaticPanelOfNormals}.}
}
\description{
Construct a somatic Panel of Normals according to GATK best practices.
}
\details{
\code{gatk_make_pon} is a convenience function that creates the Panel of Normals out of one or
multiple \code{bam} files according to GATK best practices. The input \code{bam} files are first mapped
using the \code{gatk_Mutect2} in a tumour-only mode, the resulting \code{vcf} files are merged into
a database using the \code{gatk_GenomicDBIImport} and then a the Panel of Normals is created using the
\code{gatk_CreateSomaticPanelOfNormals}.

\code{gatk_GenomicDBIImport} merges multiple variant call files (\code{vcf}), such as those created by
\code{gatk_Mutect2}, into a database. This database can be used to create a Panel of Normals using
the \code{gatk_CreateSomaticPanelOfNormals} call.

\code{gatk_CreateSomaticPanelOfNormals} creates a Panel of Normals \code{vcf} out of a provided database.
Panel of Normals are typically created out of multiple guranteed normal samples that use the same
same sequencing technology as the tumour sample. Panel of Normals are used in the \code{gatk_Mutect2}
to inform the caller about technological errors and properties and also to filter out
pre-existing variants.
}
