% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_prepare.r
\name{gatk_prepare}
\alias{gatk_prepare}
\title{Prepare sam/bam file according to GATK}
\usage{
gatk_prepare(
  input,
  output,
  reference,
  vcf,
  barcodes = NULL,
  outdir = NULL,
  remake = FALSE
)
}
\arguments{
\item{input}{an input sam/bam file}

\item{output}{an output sam/bam file}

\item{reference}{a reference fasta (\code{.fas}) file to which the sam/bam file was mapped}

\item{vcf}{a vcf file with known polymorphic sites}

\item{barcodes}{\strong{optional} file with cell barcodes to preserve}

\item{outdir}{\strong{optional} an output directory for intermediate files}

\item{remake}{whether to remake the output if it already exists}
}
\description{
Prepare sam/bam file from scRNAseq according to GATK best practices for further analysis,
such as the SNV detection.
}
\details{
This is a convenience function that chains individual \link{GATK} calls to prepare a mapped
sam/bam file for further SNV detection. Following steps are performed:
\itemize{
\item \strong{optional} \code{FilterSamReadsTag}-- a  barcode fitering step to preserve only high-quality cells
\item \code{SortSam} -- sort reads according to coordinates
\item \code{SplitNCigarReads} -- split reads around the N position in their CIGAR tag (see the GATK
documentation for more information)
\item \code{BaseRecalibrator} -- recalibrates the base quality scores after the \code{SplitNCigarReads}
\item \code{ApplyBQSR} -- applies the recalibrated base quality scores from \code{BaseRecalibrator}
}

Note that in contrast to the GATK best practices for the RNAseq, the \code{MarkDuplicates}
step is not performed as in scRNAseq, all sequences are barcoded.
}
\seealso{
\link{GATK} and \link{GATKR6} for a binding to individual GATK functions,
\link{gatk_snv} for another convenience function build on GATK calls
}
