% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_snv.r
\name{gatk_snv}
\alias{gatk_snv}
\title{Identify and filter SNVs}
\usage{
gatk_snv(
  bam,
  reference,
  vcf,
  normal = NULL,
  pon = NULL,
  germline = NULL,
  ps = NULL,
  outdir = NULL,
  remake = FALSE
)
}
\arguments{
\item{bam}{a sam/bam file prepared for variant calling}

\item{reference}{a reference fasta (\code{.fas}) file to which the sam/bam file was mapped}

\item{vcf}{an output gzipped \code{vcf} file containing only SNVs}

\item{normal}{\strong{optional} matched normal samples, either \code{sam/bam} files or normal sample names.
If \code{sam} or \code{bam} files are provided, the sample names are retrieved from the files using the
\code{gatk_GetSampleName}.}

\item{pon}{\strong{optional} Panel of Normals, see \code{\link{gatk_make_pon}}}

\item{germline}{\strong{optional}  a vcf file of germline population contaning allelic fractions}

\item{ps}{\strong{optional} a parameter string passed to GATK Mutect2 call}

\item{outdir}{\strong{optional} an output directory for intermediate files}

\item{remake}{\strong{optional} remake the output if it already exists}
}
\description{
Identify and filter single nucleotide variations in an input bam file.
}
\details{
This is a convenience function that chains individual \link{GATK} calls to identify and filter
SNVs. Following steps are performed:
\itemize{
\item \code{Mutect2} -- calls variant using the GATK Mutect2 variant caller
\item \code{FilterMutectalls} -- filter the raw output from Mutect2
\item \code{vcftools} -- a non GATK call to preserve only single nucleotide variants
}
}
\seealso{
\link{GATK} and \link{GATKR6} for a binding to indiviual GATK functions,
\link{gatk_prepare} and \link{gatk_make_pon} for other convenience functions build on GATK calls,
\code{\link[=vcftools_filter]{vcftools_filter()}} for more information on the \code{vcftools} call used in this function
}
