% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatkR6.r
\name{GATKR6}
\alias{GATKR6}
\title{GATK binding as an R6 class}
\description{
A GATK binding in an object-oriented form as a R6 class.
}
\details{
The GATK is predominantly used in a pipeline by chaining individual GATK function calls
to clean a bam file and prepare it for further analysis. This GATK class utilize
the R6 method chaining to facilitate this usage.
}
\examples{
\dontrun{
bam = GATKR6$new("foo.bam", "bar.fas", "baz.vcf")
bam$SortSam()$SplitNCigarReads()$Recalibrate()
bam$bam
}

}
\seealso{
\link{GATK} a group of functions that partially bind to the GATK toolkit,
\link{gatk_prepare} and \link{gatk_snv} for a convenience functions utilizing the GATK calls
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bam}}{a path to a current bam file. Modified during each method call to point to
the result of the last GATK action.}

\item{\code{original}}{a path to the original (unmodified) bam file}

\item{\code{reference}}{a path to reference genome fasta (\code{.fas}) file}

\item{\code{vcf}}{a path to the Variant Coding File (VCF)}

\item{\code{outdir}}{an output directory for individual methods}

\item{\code{remake}}{whether to remake already existing files}

\item{\code{history}}{a method call history}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GATKR6$new()}}
\item \href{#method-SortSam}{\code{GATKR6$SortSam()}}
\item \href{#method-MarkDuplicates}{\code{GATKR6$MarkDuplicates()}}
\item \href{#method-SplitNCigarReads}{\code{GATKR6$SplitNCigarReads()}}
\item \href{#method-Recalibrate}{\code{GATKR6$Recalibrate()}}
\item \href{#method-FilterSamReadsTag}{\code{GATKR6$FilterSamReadsTag()}}
\item \href{#method-clone}{\code{GATKR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new GATK object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GATKR6$new(bam, reference, vcf, outdir = NULL, remake = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bam}}{an input bam file}

\item{\code{reference}}{a reference fasta (\code{.fas}) file to which the sam/bam file was mapped}

\item{\code{vcf}}{a vcf file with known polymorphic sites}

\item{\code{outdir}}{\strong{optional} an output directory}

\item{\code{remake}}{\strong{optional} remake the output if it already exists}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{GATK} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortSam"></a>}}
\if{latex}{\out{\hypertarget{method-SortSam}{}}}
\subsection{Method \code{SortSam()}}{
Sort the SAM/BAM file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GATKR6$SortSam(output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{\strong{optional} a path for an output bam file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkDuplicates"></a>}}
\if{latex}{\out{\hypertarget{method-MarkDuplicates}{}}}
\subsection{Method \code{MarkDuplicates()}}{
Mark duplicates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GATKR6$MarkDuplicates(output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{\strong{optional} a path for an output bam file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SplitNCigarReads"></a>}}
\if{latex}{\out{\hypertarget{method-SplitNCigarReads}{}}}
\subsection{Method \code{SplitNCigarReads()}}{
Split the reads around the N in their CIGAR string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GATKR6$SplitNCigarReads(output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{\strong{optional} a path for an output bam file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recalibrate"></a>}}
\if{latex}{\out{\hypertarget{method-Recalibrate}{}}}
\subsection{Method \code{Recalibrate()}}{
Recalibrate the base quality score
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GATKR6$Recalibrate(output = NULL, table = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{\strong{optional} a path for an output bam file}

\item{\code{table}}{\strong{optional} an output path for a new base quality scores}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterSamReadsTag"></a>}}
\if{latex}{\out{\hypertarget{method-FilterSamReadsTag}{}}}
\subsection{Method \code{FilterSamReadsTag()}}{
Filter the sam/bam file according to tag and its values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GATKR6$FilterSamReadsTag(tag, values, output = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{a name of the tag that will be used for filtering reads}

\item{\code{values}}{one or multiple values of a chosen tag}

\item{\code{output}}{\strong{optional} a path for an output bam file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GATKR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
