% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunzip.r
\name{gunzip}
\alias{gunzip}
\title{Unzip file}
\usage{
gunzip(input, output = NULL, preserve = FALSE, overwrite = FALSE)
}
\arguments{
\item{input}{a compressed file}

\item{output}{an optional output file. If specified, the input file is preserved
irrespective of the \code{preserve} parameter and extracted to the output file.}

\item{preserve}{whether to keep the compressed file}

\item{overwrite}{overwrite existing file}
}
\description{
Extract a single gzipped file using the UNIX \code{gunzip} command.
}
\details{
Base R includes tools that can read gzipped files, write gzipped files, extract gzipped archives,
but not actually decompress single gzipped file. This function is a thin wrapper around the
UNIX \code{gunzip} command.
Currently supported extensions are \code{.gz} and \code{.z}. Other extensions will throw an error.
If the \code{output} is not specified, the extension is stripped from the file.
}
\examples{
\dontrun{
# will remove foo.txt.gz and produce foo.txt:
gunzip("foo.txt.gz")
# preserves foo.txt.gz and produce bar.txt
gunzip("foo.txt.gz", "bar.txt")
}

}
