% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.r
\name{hdi}
\alias{hdi}
\alias{hdi.density}
\alias{hdi.default}
\title{Calculate the highest density interval}
\usage{
hdi(x, size = 0.95, alpha = 1 - size)

\method{hdi}{density}(x, size = 0.95, alpha = 1 - size)

\method{hdi}{default}(x, size = 0.95, alpha = 1 - size)
}
\arguments{
\item{x}{a vector of values or an estimated density, such as the one returned by the \verb{[stats::density()]} function,
with x and y values describing the shape of the density}

\item{size}{the size of both tails or \code{1 - alpha}}

\item{alpha}{the size of both tails or \code{1 - interval}}
}
\value{
upper and lower bounds for a selected HDI interval
}
\description{
Calculate an empirical highest density interval (HDI).
}
\details{
The HDI is the shortest interval that contains the required proportion of values.
In general, HDI can be split into several subintervals if the underlying distribution
is multimodal. This function however provide only continuous HDI and is correct only for unimodal
distributions.

The HDI intervals, unlike the more common equal-tailed intervals,
are not invariant under non-linear transformations.
}
\examples{
x = rnorm(100)
hdi(x, alpha=0.05) # the 95\% HDI

# Alternatively, use kernel-density estimation to estimate the empirical density distribution
dens = density(x)
hdi(dens, alpha=0.05) # the 95\% HDI
hdi(dens, alpha=0.5) # the 50\% HDI

}
