% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexed_vector.r
\name{IndexedVector}
\alias{IndexedVector}
\title{Indexed vector}
\description{
An indexed vector
}
\details{
A vector that remembers its original indices.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{vector values}

\item{\code{indices}}{vector indices}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IndexedVector$new()}}
\item \href{#method-subset}{\code{IndexedVector$subset()}}
\item \href{#method-min}{\code{IndexedVector$min()}}
\item \href{#method-which_min}{\code{IndexedVector$which_min()}}
\item \href{#method-substract}{\code{IndexedVector$substract()}}
\item \href{#method-len}{\code{IndexedVector$len()}}
\item \href{#method-sum}{\code{IndexedVector$sum()}}
\item \href{#method-clone}{\code{IndexedVector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{IndexedVector} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$new(values, indices = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{a vector of values}

\item{\code{indices}}{\strong{optional} vector of indices. If not provided, \code{seq_along(values)} is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{IndexedVector} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\if{latex}{\out{\hypertarget{method-subset}{}}}
\subsection{Method \code{subset()}}{
Subset the vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$subset(indices)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{indices specifying elements to extract}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This is a standard subsetting, i.e., \code{1} returns the first element of the vector, not the one
belonging to the first index.

Note that since \code{IndexedVector} is an R6 object, this modifies the vector in place.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-min"></a>}}
\if{latex}{\out{\hypertarget{method-min}{}}}
\subsection{Method \code{min()}}{
A minimum of vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$min()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Method that calculate a minimum of vector values
}

\subsection{Returns}{
minimum of vector values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-which_min"></a>}}
\if{latex}{\out{\hypertarget{method-which_min}{}}}
\subsection{Method \code{which_min()}}{
Find indices of minimum values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$which_min(minimum = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minimum}}{\strong{optional} min a precalculated minimum.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Find which values are minimum values. Returns vector current (not original) indices
}

\subsection{Returns}{
current indices that have the minimum value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-substract"></a>}}
\if{latex}{\out{\hypertarget{method-substract}{}}}
\subsection{Method \code{substract()}}{
Substract values from vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$substract(vec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{that is substracted from IndexedVector values}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Substract values from vector, equivalent to calling \code{IndexedVector$values - vec}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-len"></a>}}
\if{latex}{\out{\hypertarget{method-len}{}}}
\subsection{Method \code{len()}}{
Get length of the vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$len()}\if{html}{\out{</div>}}
}

\subsection{Details}{
equivalent to calling \code{length(IndexedVector$values)}
}

\subsection{Returns}{
length of the IndexedVector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sum"></a>}}
\if{latex}{\out{\hypertarget{method-sum}{}}}
\subsection{Method \code{sum()}}{
Sum of the vector values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$sum()}\if{html}{\out{</div>}}
}

\subsection{Details}{
equivalent to calling \code{sum(IndexedVector$values)}
}

\subsection{Returns}{
the sum of the IndexedVector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndexedVector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
