% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is_nn}
\alias{is_nn}
\title{Check if value is NULL or NA}
\usage{
is_nn(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} if the value is \code{NULL} or all values are \code{NA}
}
\description{
Returns \code{TRUE} if object is either \code{NULL} or \code{NA}.
}
\details{
The no value \code{NULL} and missing value \code{NA} are both used to signify no value.
NULL is often used as an empty parameter in functions and \code{NA} is used for
the same reason in the \link[argparser:argparser]{argparser::argparser} package (because setting value to \code{NULL}
in list and environments means to delete it).

For this reason, a missing (unspecified) parameter can can reach function
as both \code{NULL} and \code{NA}. This is simple shorthand to test both conditions.
}
\examples{
is_nn(NULL)
is_nn(NA)
}
