% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.r
\name{is_sequential}
\alias{is_sequential}
\title{Test if values are sequential}
\usage{
is_sequential(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a TRUE/FALSE value
}
\description{
Values are sequential only if differences between each value is always exactly one.
}
