% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{missing_to_na}
\alias{missing_to_na}
\title{Convert a value to NA}
\usage{
missing_to_na(x, missing)
}
\arguments{
\item{x}{a vector, matrix or a data frame}

\item{missing}{a non-standard missing value}
}
\value{
an object where a non-standard missing value is replaced by \code{NA}
}
\description{
Convert a value considered to be an missing value to the standard \code{NA}
}
\examples{
x = c("foo", "bar", "missing")
phyloRNA:::missing_to_na(x, "missing")
}
