% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mkdir}
\alias{mkdir}
\title{Create directory}
\usage{
mkdir(dir)
}
\arguments{
\item{dir}{a directory to be created}
}
\description{
A simple utility function that creates directory recursively. This simulates
the behaviour of the Unix command \code{mkdir -p}.
If directory exists, no error or warning is reported, unlike with the base
function \code{dir.create} which throws an error.
}
