% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{phyloRNA}
\alias{phyloRNA}
\title{phyloRNA: A package for phylogenetical analysis of RNA data}
\description{
A package with functions to facilitate creating pipelines for the phylogenetic analysis
of RNA, especially single-cell RNA, data.
}
\section{External software}{

The phyloRNA package performs a number of calls to external programs. By default, the package
assumes that the path to the binaries is in the PATH environment variable and so that calling
only the name of the program is required (such as \code{samtools} instead of \code{path/to/samtools}).
If this is not true or the binary or symlink to it exists under a different name, you might
need to modify following \code{\link[=options]{options()}}:
\itemize{
\item \code{phyloRNA_cellranger} a path to the 10x Genomic's cellranger software
\item \code{phyloRNA_gatk} a path to the Broad Institute Genome Analysis Toolkit (GATK)
}
}

