% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.r
\name{prepare}
\alias{prepare}
\title{Prepare sam/bam file}
\usage{
prepare(
  input,
  output,
  reference,
  vcf,
  barcodes = NULL,
  outdir = NULL,
  remake = FALSE
)
}
\arguments{
\item{input}{an input sam/bam file}

\item{output}{an output sam/bam file}

\item{reference}{a reference fasta (\code{.fas}) file to which the sam/bam file was mapped}

\item{vcf}{a vcf file with known polymorphic sites}

\item{barcodes}{\strong{optional} file with cell barcodes to preserve}

\item{outdir}{\strong{optional} an output directory for intermediate files}

\item{remake}{whether to remake the output if it already exists}
}
\description{
Prepare sam/bam file from scRNAseq for further analysis, such as the SNV detection.
}
