% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_string.r
\name{random_string}
\alias{random_string}
\title{Create a random string}
\usage{
random_string(x, letters = NULL, numbers = NULL)
}
\arguments{
\item{x}{a length of the random string}

\item{letters}{a number of letters}

\item{numbers}{a number of numerals}
}
\value{
a random string of length \code{x} from letters and numbers
}
\description{
Create a random string from letters and numbers
}
