% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{read_column}
\alias{read_column}
\title{Read a column from a delimited file}
\usage{
read_column(x, k = 1L, sep = " ")
}
\arguments{
\item{x}{a path to a file}

\item{k}{\strong{optional} a integer value of a column that will be returned}

\item{sep}{\strong{optional} a separator that separate the columns}
}
\value{
a character vector
}
\description{
Read a k-th column from a delimited file.
}
