% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.r
\name{read_fasta}
\alias{read_fasta}
\title{Read fasta file}
\usage{
read_fasta(file)
}
\arguments{
\item{file}{a fasta file}
}
\value{
a named vector of sequences
}
\description{
Read a fasta file and return a named vector of sequences.
}
\examples{
# Construct sequences as a named vector
seq1 = as.character(1:5)
names(seq1) = letters[1:5]

# Write sequences in a fasta format
fasta = tempfile()
writeLines(paste0(">", names(seq1), "\n", seq1), fasta)
# or alternatively, write_fasta(seq1)

# Read sequences with read_fasta
seq2 = read_fasta(fasta)
identical(seq1, seq2)

}
\seealso{
\link{write_fasta} to write fasta file,
\link{table2fasta} to transform table into a fasta file,
\link{fasta2table} tp transform fasta into a table
}
