% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{remap}
\alias{remap}
\title{Remap BAM file to a new reference genome}
\usage{
remap(
  input,
  reference,
  annotation,
  outdir,
  fastqdir = NULL,
  refdir = NULL,
  countdir = NULL,
  nthreads = 4,
  remake = FALSE,
  cbam = FALSE,
  cbar = FALSE
)
}
\arguments{
\item{input}{a bam file previously mapped by the Cellranger software}

\item{reference}{a reference genome fasta file (e.g., \code{.fna})}

\item{annotation}{a GTF (\code{.gtf}) annotation file associated with the reference genome}

\item{outdir}{a directory where the}

\item{fastqdir}{an optional directory for the \link{bamtofastq}, see details}

\item{refdir}{an optional directory for the \link{cellranger_mkref}, see details}

\item{countdir}{an optional directory for the \link{cellranger_count}, see details}

\item{nthreads}{a number of threads to run on}

\item{remake}{whether to remake the output if it already exists}

\item{cbam}{a TRUE value or file path to a location where the re-mapped bam file will be copied}

\item{cbar}{a TRUE value or file path to a location where the file with filtered barcodes
will be copied}
}
\description{
This function will remap a bam file previously mapped by the Cellranger software to a new genome.
}
\details{
The \code{remap} function calls the \link{bamtofastq}, \link{cellranger_mkref} and \link{cellranger_count} to
convert the bam file into fastq files, create the cellranger reference genome and perform
the expression analysis respectively. See the individual functions for details.

The optional directories \code{fastqdir}, \code{refdir} and \code{countdir} are for outputs from the
above mentioned \link{bamtofastq}, \link{cellranger_mkref} and \link{cellranger_count}. If not specified,
folders \code{fastqdir}, \code{ref} and \code{count} in the \code{outdir} directory are used. If these are
undesired, a path to a non-existing directory or directory with a previously existing run
must be used. For example when a multiple bam files are being re-mapped to the same reference,
all of them would share the same \code{refdir}.
See the \link{bamtofastq}, \link{cellranger_mkref} and \link{cellranger_count} for details.

Given the intended usage is to obtain a remapped bam file and a list of filtered barcodes,
\code{cbam} and \code{cbar} parameters are provided. If \code{TRUE}, the bam file and barcodes are copied
to the \code{outdir} folder as \code{prefix.bam} and \code{prefix.barcodes.txt} where \code{prefix} is the core
of the input file (see \link{corename}). An arbitrary path can also be provided.
}
\seealso{
\link{bamtofastq}, \link{cellranger_mkref} and \link{cellranger_count} for details on the steps
performed by the \code{remap} function.
}
