% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_constant.r
\name{remove_constant}
\alias{remove_constant}
\title{Remove constant}
\usage{
remove_constant(data, margin = 1, unknown = "N")
}
\arguments{
\item{data}{a data matrix}

\item{margin}{\strong{optional} rows (1) or columns (2) that will be explored}

\item{unknown}{\strong{optional} elements that are excluded from this comparison
(do not count among the unique values).}
}
\value{
a data matrix without
}
\description{
Remove constant rows or columns from matrix.
}
\details{
Constant row or column is such column that has only single unique value
(beside the unknown elements).
}
