% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{replace.ordinal}
\alias{replace.ordinal}
\title{Replace missing categories in ordinal scale}
\usage{
replace.ordinal(x)
}
\arguments{
\item{x}{a vector or a matrix of categorical variables}
}
\value{
a vector or a matrix where categories are replaced with
a sequential numeric vector
}
\description{
Rescale an ordinal scale to replace missing categories.
This will assure that the categories in the ordinal scale are sequential.
}
\examples{
foo = c(1, 3, 5)
bar = c(1, 2, 3)
identical(replace.ordinal(foo), bar)

}
