% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{replace}
\alias{replace}
\alias{replace.default}
\alias{replace.data.frame}
\title{Replace values in a vector or a matrix}
\usage{
replace(x, ...)

\method{replace}{default}(x, values, replace, ...)

\method{replace}{data.frame}(x, values, replace, ...)
}
\arguments{
\item{x}{an object}

\item{...}{further arguments passed to or from other methods}

\item{values}{a vector of values in 'x' to be replaced by values in \code{replace}}

\item{replace}{replacement values}
}
\value{
an object with replaced values
}
\description{
Replace the values in \code{x}, specified by the vector of values
\code{values} with the values in the vector \code{replace}.
}
\examples{
vec = c("foo", "bar", "baz")
replace(vec, "bar", "spam")
replace(vec, c("foo", "bar"), c("bar", "foo"))
replace(vec, "spam", "eggs")
replace(vec, c("foo", "bar", "baz"), c("spam", "spam", "spam"))

}
