% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{replace_ordinal}
\alias{replace_ordinal}
\title{Replace missing categories on ordinal scale}
\usage{
replace_ordinal(x, replace = NULL)
}
\arguments{
\item{x}{an object with ordinal categories}

\item{replace}{\strong{optional} a vector of replacement values}
}
\value{
x where all categories are present and sequential
a sequential numeric vector
}
\description{
Rescale an ordinal scale to replace missing categories.
This will assure that the currently present ordinal categories are sequential.
Alternatively, user can provided replacement values. In such case,
the user-provided values will be used as-is.
}
\examples{
foo = c(1, 3, 5)
bar = c(1, 2, 3)
identical(replace_ordinal(foo), bar)

foo = c(1, 5, 3)
bar = c("a", "c", "b")
identical(replace_ordinal(foo, letters), bar)

}
