% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2seq.r
\name{seq2tab}
\alias{seq2tab}
\title{Convert sequences into a table}
\usage{
seq2tab(x, margin = 1)
}
\arguments{
\item{x}{a named vector of sequences, such as from \code{read_fasta}}

\item{margin}{\strong{optional} whether rows sequences should be in rows (1)
or columns (2)}
}
\value{
a matrix of fasta sequences
}
\description{
Convert a named vector of sequences into a matrix by splitting up sequences
into individual positions.
}
\examples{
seq = c("ac", "bd")
mat = seq2tab(seq)
identical(seq, c("ac","bd"))

}
\seealso{
\link{tab2seq} to transform table into a vector of sequences,
\link{read_fasta} to read sequences from a fasta file,
\link{write_fasta} to write sequences into a fasta file
}
