% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemE.r
\name{systemE}
\alias{systemE}
\title{Run an external command}
\usage{
systemE(command, args, call = FALSE, dir = NA)
}
\arguments{
\item{command}{an external command to run}

\item{args}{an arguments to the command}

\item{call}{prints the command call, useful for debugging}

\item{dir}{run the command in a target directory. Some external commands create output
in the directory they are executed. This is often less than ideal and leads to an unwanted
shuffling with files or increased complexity to accomodate to this behaviour.}
}
\description{
A wrapper around the system2 function that raise an error if the external
external command resulted in an error. See the \code{link{system2}} function
for more information.
}
