% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2seq.r
\name{tab2seq}
\alias{tab2seq}
\title{Convert a table into a vector of sequences}
\usage{
tab2seq(x, margin = 1, na = c("N", "-", "?"))
}
\arguments{
\item{x}{data.frame or matrix, symbols must have length 1}

\item{margin}{\strong{optional} whether rows (1) or columns (2) should be
concatenated}

\item{na}{\strong{optional} substitute \code{NA} for a missing/unknown symbol.}
}
\value{
a vector of sequences
}
\description{
Convert a data.frame or matrix into a vector of sequences.
This is done by concatenating elements either row-wise (margin 1)
or column-wise (margin 2).
}
\examples{
mat = matrix(letters[1:4],2,2)
fasta = tab2seq(mat)
identical(fasta, c("ac", "bd"))

}
\seealso{
\link{seq2tab} to transform sequences into a table,
\link{read_fasta} to read sequences from a fasta file,
\link{write_fasta} to write sequences into a fasta file
}
