% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.r
\name{table2fasta}
\alias{table2fasta}
\title{Convert a table into fasta format}
\usage{
table2fasta(x, margin = 1, na = c("N", "-", "?"))
}
\arguments{
\item{x}{data.frame or matrix, symbols must have length 1}

\item{margin}{\strong{optional} whether rows (1) or columns (2) should be concatenated}

\item{na}{substitute \code{NA} for a missing/unknown symbol.}
}
\value{
a vector of fasta files
}
\description{
Convert a data.frame or matrix into a fasta format, named vector of sequences.
This is done by concatenating elements either row-wise (margin 1) or column-wise (margin 2).
}
\examples{
mat = matrix(letters[1:4],2,2)
fasta = table2fasta(mat)
identical(fasta, c("ac", "bd"))

}
\seealso{
\link{write_fasta} to write fasta file,
\link{read_fasta} to read fasta file,
\link{fasta2table} tp transform fasta into a table
}
