% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vff.r
\name{vff_make}
\alias{vff_make}
\title{Make a varian frequency files}
\usage{
vff_make(bam, vcf, barcodes = NULL, folder = NULL)
}
\arguments{
\item{bam}{an input bam file}

\item{vcf}{an input vcf file}

\item{barcodes}{\strong{optional} a barcode file}

\item{folder}{\strong{optional} an output folder}
}
\description{
Calculate the base frequency for every cell barcode in the bam file for every variant in the vcf.
}
\details{
This code is a thin wrapper over the Python script \code{vff_make} found in the package's \code{inst}
folder.
}
