% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vff.r
\name{vff_merge}
\alias{vff_merge}
\title{Merge variant frequency files}
\usage{
vff_merge(folder, output)
}
\arguments{
\item{folder}{a folder with the variant frequency files}

\item{output}{a path to an output table}
}
\description{
Merge the variatn frequency files previously created by the \link{vff_make} and outputs an alignment
table with the most frequent base for every cell at each position.
}
\details{
This code is a thin wrapper over the Python script \code{vff_merge.py} found in the package's \code{inst}
folder.
}
