% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.r
\name{write_fasta}
\alias{write_fasta}
\title{Write fasta file}
\usage{
write_fasta(seq, file)
}
\arguments{
\item{seq}{a named vector of sequences}

\item{file}{an output file}
}
\description{
Write a named vector of sequences to a file in a fasta format.
}
\examples{
# Construct sequences as a named vector
seq = as.character(1:5)
names(seq) = letters[1:5]

# Write sequences to a temp file
fasta = tempfile()
write_fasta(seq, fasta)

identical(paste0(">", names(seq), "\n", seq, collapse="\n"), readLines(fasta))

}
\seealso{
\link{read_fasta} to read sequences from a fasta file
}
