% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.r
\name{expr}
\alias{expr}
\alias{expr_read10x}
\alias{expr_read10xh5}
\alias{expr_normalize}
\alias{expr_scale}
\title{Functions for manipulation with the expression data}
\usage{
expr_read10x(
  dir,
  gene_column = 2,
  unique_features = TRUE,
  strip_suffix = FALSE
)

expr_read10xh5(input, use_names = TRUE, unique_features = TRUE)

expr_normalize(data, scale_factor = 10000)

expr_scale(data)
}
\arguments{
\item{dir}{a directory with barcodes, features and sparse matrix}

\item{gene_column}{\strong{optional} the position of column with gene/feature names}

\item{unique_features}{\strong{optional} gene/feature names will be made unique to prevent possible
name conflict}

\item{strip_suffix}{\strong{optional} the \code{-1} suffix which is common for 10X barcodes}

\item{input}{an input data in the \code{.h5} format}

\item{use_names}{\strong{optional} use gene names instead of gene IDs}

\item{data}{an expression matrix}

\item{scale_factor}{\strong{optional} a scaling factor}
}
\value{
sparse matrix

a list of sparse matrices

log-normalized matrix

rescaled and centered data
}
\description{
A group of functions, often lifted and modified from the Seurat package for manipulation with the 10X scRNAseq data.
}
\details{
The Seurat package is a great tool for manipulation with the 10X scRNAseq expression data. However, it has two major issues. The first one is that it assumes that the zero expression is true zero. While this is reasonable assumption with a high coverage, small coverage scRNAseq can suffer for drop out due to the nature of a small amount of starting product and certain randomness coming from used methodology. This means that the measured zero level of expression is more accurately described as a missing data. Unfortunatelly, the sparse matrice implementation used by Seurat can not allow this change of context.

The second issue is the huge amount of dependencies that the Seurat brings. Due to the limited scope in which Seurat functionality is used and given that the utilized functionality had to be already rewritten due to the above reasons, it seems more convenient to just lift up remaining Seurat functionality.
}
\section{Functions}{
\itemize{
\item \code{expr_read10x}: Read 10X data

\item \code{expr_read10xh5}: Read 10X data in the \code{.h5} format.

\item \code{expr_normalize}: Log-normalize data. Feature counts for each cell are divided by the total count
for that cell multiplied by a scale factor. This is then natural log transformed using log1p.

\item \code{expr_scale}: Scale and center features
}}

