% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.r
\name{GATK}
\alias{GATK}
\alias{gatk_MarkDuplicates}
\alias{gatk_SortSam}
\alias{gatk_SplitNCigarReads}
\alias{gatk_BaseRecalibrator}
\alias{gatk_ApplyBQSR}
\alias{gatk_IndexFeatureFile}
\alias{gatk_FilterSamReadsTag}
\title{Mapping to the Genome Analysis ToolKit (GATK)}
\usage{
gatk_MarkDuplicates(input, output, remake = FALSE)

gatk_SortSam(input, output, remake = FALSE)

gatk_SplitNCigarReads(input, output, reference, remake = FALSE)

gatk_BaseRecalibrator(input, reference, vcf, table, remake = FALSE)

gatk_ApplyBQSR(input, reference, table, output, remake = FALSE)

gatk_IndexFeatureFile(vcf, remake = FALSE)

gatk_FilterSamReadsTag(input, output, tag, values, remake = FALSE)
}
\arguments{
\item{input}{an input BAM file}

\item{output}{an output BAM file}

\item{remake}{whether to remake the output if it already exists}

\item{reference}{a reference fasta (\code{.fas}) file to which the bam file was mapped}

\item{vcf}{a vcf file with known polymorphic sites}

\item{table}{a table with recalibration information}

\item{tag}{a name of tag}

\item{values}{one or multiple values of particular tag to keep}
}
\description{
Group of function that map to a functionality of Broadinstitute's Genome Analysis ToolKit.
}
\section{Functions}{
\itemize{
\item \code{gatk_MarkDuplicates}: Mark duplicate reads in SAM/BAM file. These reads are then marked with a SAM tag

\item \code{gatk_SortSam}: Sort SAM or BAM file according to coordinates.

\item \code{gatk_SplitNCigarReads}: Split reads that contains N in their CIGAR string

\item \code{gatk_BaseRecalibrator}: Recalibrate the base quality score and outputs a table of new recalibrated
values

\item \code{gatk_ApplyBQSR}: Apply the base quality score recalibration according to the recalculated
scores from the \verb{[gatk_BaseRecalibrator]}

\item \code{gatk_IndexFeatureFile}: Create an index for the Variant Coding File

\item \code{gatk_FilterSamReadsTag}: Filter sam/bam file according to tag and its value
}}

\seealso{
\link{GATKR6} a GATK binding in the form of R6 class
}
