% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.r
\name{gunzip}
\alias{gunzip}
\title{Unzip file}
\usage{
gunzip(file, preserve = FALSE, overwrite = FALSE)
}
\arguments{
\item{file}{a compressed file}

\item{preserve}{whether to keep the compressed file}

\item{overwrite}{overwrite existing file}
}
\description{
Extract a single gzipped file using the UNIX \code{gunzip} command.
}
\details{
Base R includes tools that can read gzipped files, write gzipped files, extract gzipped archives,
but not actually decompress single gzipped file.
}
