% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.r
\name{hdi}
\alias{hdi}
\title{Calculate the highest density interval}
\usage{
hdi(dens, alpha = 0.05)
}
\arguments{
\item{dens}{an estimated density, such as the one returned by the \verb{[stats::density()]} function,
with x and y values describing the shape of the density}

\item{alpha}{the size of both tails or \code{1 - interval}}
}
\value{
upper and lower bounds for a selected HDI interval
}
\description{
Calculate the highest density interval (HDI).
}
\details{
The HDI is the shortest interval with the required density.
Note that in general, HDI can be split into several subintervals if te underlying distribution
is multimodal. This function however provide only continuous HDI.

Also note that the HDI intervals, unlike the more common equal-tailed intervals,
are not invariant under non-linear transformations.
}
\examples{
x = rnorm(100)
dens = density(x)
hdi(dens, alpha=0.05) # the 95\% HDI
hdi(dens, alpha=0.5) # the 50\% HDI

}
