% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{replace}
\alias{replace}
\title{Replace values in a vector or a matrix}
\usage{
replace(x, values, replace)
}
\arguments{
\item{x}{a vector or a matrix}

\item{values}{a vector of values in 'x' to be replaced by values in \code{replace}}

\item{replace}{replacement values}
}
\value{
a vector a a matrix with replaced values
}
\description{
Replace the values in a vector or a matrix \code{x}, specified by the vector of values
\code{values} with the values in the vector \code{replace}.
}
\examples{
vec = c("foo", "bar", "baz")
replace(vec, "bar", "spam")
replace(vec, c("foo", "bar"), c("bar", "foo"))
replace(vec, "spam", "eggs")
replace(vec, c("foo", "bar", "baz"), c("spam", "spam", "spam"))

}
